requirements-production.txt
# Production Requirements for Job Genie - Enhanced Resume Matcher
# Generated: December 2024
# Python 3.10+

# Core Framework
streamlit>=1.28.1,<2.0.0
streamlit-option-menu>=0.3.6

# Data Processing & Analytics
pandas>=2.1.0,<3.0.0
numpy>=1.24.0,<2.0.0
plotly>=5.17.0
altair>=5.1.0

# NLP & AI Models
spacy>=3.7.0,<4.0.0
transformers>=4.35.0,<5.0.0
sentence-transformers>=2.2.2
torch>=2.0.0,<3.0.0
tokenizers>=0.14.0

# Document Processing
PyPDF2>=3.0.1
python-docx>=1.1.0
openpyxl>=3.1.2
python-magic>=0.4.27

# Vector Database & Embeddings
qdrant-client>=1.6.0
fastembed>=0.1.3
cohere>=4.32.0

# HTTP & API
requests>=2.31.0
httpx>=0.25.0
aiohttp>=3.9.0

# Data Validation & Security
pydantic>=2.4.0
pydantic-settings>=2.0.3
cryptography>=41.0.7
python-dotenv>=1.0.0

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3
uuid>=1.30
pathlib>=1.0.1

# Development Tools (Optional)
black>=23.9.1
flake8>=6.1.0
mypy>=1.6.0
pytest>=7.4.2
pytest-cov>=4.1.0
pre-commit>=3.5.0

# Web & Deployment
gunicorn>=21.2.0
uvicorn>=0.24.0

# File Handling
chardet>=5.2.0
python-mimeparse>=1.6.0

# Configuration
PyYAML>=6.0.1
toml>=0.10.2

# Async Support
asyncio>=3.4.3
aiofiles>=23.2.1

# Database
sqlite3>=2.6.0

# Visualization Extras
matplotlib>=3.7.2
seaborn>=0.12.2

# Performance
cachetools>=5.3.2
diskcache>=5.6.3

# Monitoring & Logging
structlog>=23.2.0

# Cloud Storage (Optional)
boto3>=1.28.0
azure-storage-blob>=12.19.0

# Email (Optional)
smtplib>=3.0.0

# Progress Bars
tqdm>=4.66.0

# Rich Text Processing
rich>=13.6.0
textstat>=0.7.3

# Machine Learning Utilities
scikit-learn>=1.3.0
joblib>=1.3.2

# Language Detection
langdetect>=1.0.9

# Resume Parsing Enhancement
python-resume-parser>=0.8.4

# Caching
redis>=5.0.0

# Task Queue (Optional)
celery>=5.3.0

# API Documentation
fastapi>=0.104.1
uvicorn[standard]>=0.24.0

# Testing Utilities
factory-boy>=3.3.0
faker>=19.12.0

# Environment Management
python-decouple>=3.8

# File Format Support
xlsxwriter>=3.1.9
python-pptx>=0.6.21

# Text Processing
nltk>=3.8.1
regex>=2023.10.3

# Dependency Management
pip-tools>=7.3.0

# Jupyter Support (Optional)
jupyter>=1.0.0
ipywidgets>=8.1.1

# Web Scraping (Optional)
beautifulsoup4>=4.12.2
selenium>=4.15.0

# Advanced NLP
gensim>=4.3.2
wordcloud>=1.9.2

# Database ORM (Optional)
sqlalchemy>=2.0.23
alembic>=1.12.1

# Image Processing (Optional)
Pillow>=10.0.1

# Financial/Date Parsing
python-dateutil>=2.8.2
dateparser>=1.1.8

# Configuration Validation
cerberus>=1.3.4

# Advanced Analytics
scipy>=1.11.3
statsmodels>=0.14.0
